録画完了後の処理を Operator から RecordingFinModel へ移動
